+++ src/lib/shared/alert-dialog/sz-alert-dialog.component.ts
@@ -1,0 +1,35 @@
+import { Component, Inject } from '@angular/core';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+
+@Component({
+    selector: 'sz-alert-dialog',
+    templateUrl: 'sz-alert-dialog.component.html',
+    styleUrls: ['sz-alert-dialog.component.scss']
+  })
+  export class SzAlertMessageDialog {
+    private _showOkButton = true;
+    public title: string = 'Alert';
+    public text: string;
+    public buttonText: string = "Ok";
+
+    public get showDialogActions(): boolean {
+      return this._showOkButton;
+    }
+  
+    constructor(@Inject(MAT_DIALOG_DATA) public data: { title?: string, text?: string, buttonText?: string, showButton?: boolean }) {
+      if(data) {
+        if(data.title) {
+          this.title = data.title;
+        }
+        if(data.text) {
+          this.text = data.text;
+        }
+        if(data.buttonText) {
+          this.buttonText = data.buttonText;
+        }
+        if(data.showButton) {
+          this._showOkButton = data.showButton;
+        }
+      }
+    }
+  }